requirements.txt
pytube==9.6.0
pytube3==9.6.4
typing-extensions==3.7.4.2

Pillow~=7.2.0